+++ Src/Interpreter/Moe/Syscall/Moe.Syscall.cs
@@ -13,7 +13,7 @@ public partial class MoeInterpreter
 		{
 			string s = "";
 			s += $"\t{variable}";
-			if (variable.Type == MoeVariableType.Int || variable.Type == MoeVariableType.Double)
+			if (variable.Type == MoeVariableType.Int || variable.Type == MoeVariableType.Float)
 				for (int i = 0; i < variable.Size; i++)
 					s += $"{(i % 5 == 0 ? "\n" : "")}\t\tobj[{i}]: {variable[i]}";
 			else if (variable.Type == MoeVariableType.String)
