+++ Src/Types/DataStructure.cs
@@ -60,22 +60,22 @@ public record class FRect
 	}
 
 	public static implicit operator SKRectI(FRect rect) => new((int)rect.Left, (int)rect.Top, (int)rect.Right, (int)rect.Bottom);
-	public static implicit operator SKRect(FRect rect) => new((float)rect.Left, (float)rect.Top, (float)rect.Right, (float)rect.Bottom);
+	public static implicit operator SKRect(FRect rect) => new(rect.Left, rect.Top, rect.Right, rect.Bottom);
 	public static explicit operator (IVector, IVector)(FRect rect) => (new((int)rect.X, (int)rect.Y), new((int)rect.H, (int)rect.W));
 	public static explicit operator (FVector, FVector)(FRect rect) => (new(rect.X, rect.Y), new(rect.H, rect.W));
 
 	public static explicit operator IRect(FRect rect) => new((int)rect.X, (int)rect.Y, (int)rect.W, (int)rect.H);
 
 
-	public double X { get; set; } = 0;
-	public double Y { get; set; } = 0;
-	public double W { get; set; } = 0;
-	public double H { get; set; } = 0;
+	public float X { get; set; } = 0;
+	public float Y { get; set; } = 0;
+	public float W { get; set; } = 0;
+	public float H { get; set; } = 0;
 
-	[JsonIgnore] public double Left => X;
-	[JsonIgnore] public double Right => X + W;
-	[JsonIgnore] public double Top => Y;
-	[JsonIgnore] public double Bottom => Y + H;
+	[JsonIgnore] public float Left => X;
+	[JsonIgnore] public float Right => X + W;
+	[JsonIgnore] public float Top => Y;
+	[JsonIgnore] public float Bottom => Y + H;
 
 	[JsonIgnore] public FVector MidPoint => new(X + W / 2, Y + H / 2);
 }
@@ -96,7 +96,7 @@ public record struct IVector(int X, int Y)
 	public static implicit operator SKSize(IVector p) => new(p.X, p.Y);
 
 	public static explicit operator (int, int)(IVector p) => (p.X, p.Y);
-	public static explicit operator (double, double)(IVector p) => (p.X, p.Y);
+	public static explicit operator (float, float)(IVector p) => (p.X, p.Y);
 
 	public static explicit operator FVector(IVector p) => new(p.X, p.Y);
 
@@ -104,28 +104,28 @@ public record struct IVector(int X, int Y)
 	[JsonIgnore] public readonly int Height => Y;
 }
 
-public record struct FVector(double X, double Y)
+public record struct FVector(float X, float Y)
 {
-	public static FVector operator *(FVector v1, double off) => new(v1.X * off, v1.Y * off);
-	public static FVector operator /(FVector v1, double off) => new(v1.X / off, v1.Y / off);
+	public static FVector operator *(FVector v1, float off) => new(v1.X * off, v1.Y * off);
+	public static FVector operator /(FVector v1, float off) => new(v1.X / off, v1.Y / off);
 	public static FVector operator *(FVector v1, FVector v2) => new(v1.X * v2.X, v1.Y * v2.Y);
 	public static FVector operator +(FVector v1, FVector v2) => new(v1.X + v2.X, v1.Y + v2.Y);
 	public static FVector operator -(FVector v1, FVector v2) => new(v1.X - v2.X, v1.Y - v2.Y);
 	public static FVector operator -(FVector v1) => new(-v1.X, -v1.Y);
 
 	public static implicit operator SKPointI(FVector p) => new((int)p.X, (int)p.Y);
-	public static implicit operator SKPoint(FVector p) => new((float)p.X, (float)p.Y);
+	public static implicit operator SKPoint(FVector p) => new(p.X, p.Y);
 
 	public static implicit operator SKSizeI(FVector p) => new((int)p.X, (int)p.Y);
-	public static implicit operator SKSize(FVector p) => new((float)p.X, (float)p.Y);
+	public static implicit operator SKSize(FVector p) => new(p.X, p.Y);
 
 	public static explicit operator (int, int)(FVector p) => ((int)p.X, (int)p.Y);
-	public static explicit operator (double, double)(FVector p) => (p.X, p.Y);
+	public static explicit operator (float, float)(FVector p) => (p.X, p.Y);
 
 	public static explicit operator IVector(FVector p) => new((int)p.X, (int)p.Y);
 
-	[JsonIgnore] public readonly double Width => X;
-	[JsonIgnore] public readonly double Height => Y;
+	[JsonIgnore] public readonly float Width => X;
+	[JsonIgnore] public readonly float Height => Y;
 }
